chore(mcp): extract duplicate mock helpers into shared test-support module
Move mock_test_config(), new_mock(), and state_with_mock() from
integration.rs and dyn_dispatch.rs into tests/common/mod.rs so both
test binaries share a single implementation. The token string is
unified to "ghp_mock_token" (no test asserts on the value).

Issue-specific builders (mock_issue, make_issue, test_issue) remain
local to each file since they differ in structure and purpose.
